LayoutTests/accessibility/language-attribute-change-expected.txt

Hola Mundo!
PASS accessibleContent.language is 'AXLanguage: en'
PASS accessibleContent.language is 'AXLanguage: es'
PASS successfullyParsed is true

TEST COMPLETE
